otus-py/requirements.txt
Flask==2.1.0
Flask-SQLAlchemy
Flask-Migrate
psycopg2==2.9.3
Flask-Script
